[txg] `TXCount` newtype
